Consolidate test coverage for maven-publish of java libraries
